Separate superfluous return mixin
